+++ .gitignore
@@ -1,2 +1,3 @@
 *csv
 *premodels/*
+*pycache*
