+++ .travis.yml
@@ -6,4 +6,4 @@ install:
   - pip install numpy>=1.16.4 keras>=2.2.5 tensorflow>=1.14.0 scikit-learn>=0.21.2 wget pandas
   - pip install -e .
 # command to run tests
-script: pytest -s -v tests.py
+script: pytest -s -v ./tests/tests.py
