+++ tests/tests.py
@@ -1,6 +1,10 @@
-from deepstack import Member
-from deepstack import DirichletEnsemble
-from deepstack import StackEnsemble
+import sys
+sys.path.append(".")
+sys.path.append("..")
+sys.path.append("../deepstack/")
+from deepstack.base import Member
+from deepstack.ensemble import DirichletEnsemble
+from deepstack.ensemble import StackEnsemble
 from sklearn.datasets.samples_generator import make_blobs
 from keras.models import Sequential
 import random
@@ -17,14 +21,16 @@ from keras.preprocessing.image import ImageDataGenerator
 from keras.layers import Conv2D, MaxPooling2D
 from keras import regularizers
 from sklearn.ensemble import RandomForestRegressor
-from keras.layers import Dense, Activation, Flatten, Dropout, BatchNormalization
+from keras.layers import Dense, Activation, Flatten, Dropout
+from keras.layers import BatchNormalization
 
 
 def _get_fitted_random_model(trainX, trainy):
     model = Sequential()
     model.add(Dense(random.randint(20, 30), input_dim=2, activation='relu'))
     model.add(Dense(1, activation='sigmoid'))
-    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
+    model.compile(loss='binary_crossentropy', optimizer='adam', 
+                  metrics=['accuracy'])
     model.fit(trainX, trainy, epochs=50, verbose=0)
     return model
 
@@ -34,7 +40,8 @@ def test_dirichletensemble():
     Tests if builsing an dirichlet ensemble is running without problems
     """
     np.random.seed(seed=2)
-    X, y = make_blobs(n_samples=200, centers=2, n_features=2, cluster_std=4, random_state=2)
+    X, y = make_blobs(n_samples=200, centers=2, n_features=2, cluster_std=4, 
+                      random_state=2)
     n_train = 100
     trainX, testX = X[:n_train, :], X[n_train:, :]
     trainy, testy = y[:n_train], y[n_train:]
@@ -44,7 +51,8 @@ def test_dirichletensemble():
         model = _get_fitted_random_model(trainX, trainy)
         train_batches = CustomIterator(trainX, trainy, 32)
         val_batches = CustomIterator(testX, testy, 32)
-        m = Member(keras_model=model, name="Model " + str(i), train_batches=train_batches, val_batches=val_batches)
+        m = Member(keras_model=model, name="Model " + str(i), 
+                   train_batches=train_batches, val_batches=val_batches)
         stack.add_member(m)
     stack.fit()
     stack.describe()
@@ -87,8 +95,10 @@ def test_stackensemble():
     df = pd.read_csv("sonar.csv", header=None)
     classes = df[60].map({"M": 0, "R": 1})
     del df[60]
-    X_train, X_test, y_train, y_test = train_test_split(df, classes, test_size=2 / 3, random_state=1)
-    X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.5, random_state=1)
+    X_train, X_test, y_train, y_test = train_test_split(
+        df, classes, test_size=2 / 3, random_state=1)
+    X_train, X_val, y_train, y_val = train_test_split(
+        X_train, y_train, test_size=0.5, random_state=1)
 
     y_train = to_categorical(y_train)
     y_val = to_categorical(y_val)
@@ -98,7 +108,8 @@ def test_stackensemble():
     model1.add(Dense(60, input_dim=60, activation='relu'))
     model1.add(Dense(30, activation='relu'))
     model1.add(Dense(2, activation='softmax'))
-    model1.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
+    model1.compile(loss='categorical_crossentropy',
+                   optimizer='adam', metrics=['accuracy'])
     train_batches1 = CustomIterator(X_train, y_train, 32)
     val_batches1 = CustomIterator(X_val, y_val, 32)
     test_batches1 = CustomIterator(X_test, y_test, 32)
@@ -109,13 +120,15 @@ def test_stackensemble():
         validation_steps=val_batches1.n // 32,
         epochs=10
     )
-    member1 = Member(keras_model=model1, train_batches=val_batches1, val_batches=test_batches1, name="Model1")
+    member1 = Member(keras_model=model1, train_batches=val_batches1, 
+                     val_batches=test_batches1, name="Model1")
 
     model2 = Sequential()
     model2.add(Dense(15, input_dim=60, activation='relu'))
     model2.add(Dense(7, activation='relu'))
     model2.add(Dense(2, activation='softmax'))
-    model2.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
+    model2.compile(loss='categorical_crossentropy',
+                   optimizer='adam', metrics=['accuracy'])
     train_batches2 = CustomIterator(X_train, y_train, 32)
     val_batches2 = CustomIterator(X_val, y_val, 32)
     test_batches2 = CustomIterator(X_test, y_test, 32)
@@ -126,7 +139,8 @@ def test_stackensemble():
         validation_steps=val_batches2.n // 32,
         epochs=10
     )
-    member2 = Member(keras_model=model2, train_batches=val_batches2, val_batches=test_batches2, name="Model2")
+    member2 = Member(keras_model=model2, train_batches=val_batches2,
+                     val_batches=test_batches2, name="Model2")
 
     if not os.path.exists("./premodels/"):
         os.mkdir("./premodels/")
@@ -147,7 +161,7 @@ def test_stackensemble():
 
     if platform == "darwin":
         print(auc1, auc2)
-        assert(auc1 == auc2)   # TODO: No idea why this is not working under linux
+        assert(auc1 == auc2)  # TODO: not working under linux
 
     member1.load_kerasmodel("./premodels/model1.h5")
     member2.load_kerasmodel("./premodels/model2.h5")
@@ -165,7 +179,7 @@ def test_stackensemble():
 
     if platform == "darwin":
         print(auc3, auc4)
-        assert(auc3 == auc4)   # TODO: No idea why this is not working under linux
+        assert(auc3 == auc4)   # TODO: not working under linux
 
 
 def _load_cifar_dataset(trainsample=5000, testsample=500):
@@ -182,29 +196,35 @@ def _create_random_cifar_model(input_shape):
     weight_decay = 1e-4
     num_classes = 10
     model = Sequential()
-    model.add(Conv2D(32, (3, 3), padding='same', kernel_regularizer=regularizers.l2(weight_decay),
+    model.add(Conv2D(32, (3, 3), padding='same',
+                     kernel_regularizer=regularizers.l2(weight_decay),
                      input_shape=input_shape))
     model.add(Activation('elu'))
     model.add(BatchNormalization())
-    model.add(Conv2D(random.randint(16, 64), (3, 3), padding='same', kernel_regularizer=regularizers.l2(weight_decay)))
+    model.add(Conv2D(random.randint(16, 64), (3, 3), padding='same', 
+                     kernel_regularizer=regularizers.l2(weight_decay)))
     model.add(Activation('elu'))
     model.add(BatchNormalization())
     model.add(MaxPooling2D(pool_size=(2, 2)))
     model.add(Dropout(random.randint(0, 5) * 0.1))
 
-    model.add(Conv2D(random.randint(16, 64), (3, 3), padding='same', kernel_regularizer=regularizers.l2(weight_decay)))
+    model.add(Conv2D(random.randint(16, 64), (3, 3), padding='same',
+                     kernel_regularizer=regularizers.l2(weight_decay)))
     model.add(Activation('elu'))
     model.add(BatchNormalization())
-    model.add(Conv2D(random.randint(16, 64), (3, 3), padding='same', kernel_regularizer=regularizers.l2(weight_decay)))
+    model.add(Conv2D(random.randint(16, 64), (3, 3), padding='same',
+                     kernel_regularizer=regularizers.l2(weight_decay)))
     model.add(Activation('elu'))
     model.add(BatchNormalization())
     model.add(MaxPooling2D(pool_size=(2, 2)))
     model.add(Dropout(random.randint(0, 5) * 0.1))
 
-    model.add(Conv2D(random.randint(64, 128), (3, 3), padding='same', kernel_regularizer=regularizers.l2(weight_decay)))
+    model.add(Conv2D(random.randint(64, 128), (3, 3), padding='same',
+                     kernel_regularizer=regularizers.l2(weight_decay)))
     model.add(Activation('elu'))
     model.add(BatchNormalization())
-    model.add(Conv2D(random.randint(64, 128), (3, 3), padding='same', kernel_regularizer=regularizers.l2(weight_decay)))
+    model.add(Conv2D(random.randint(64, 128), (3, 3), padding='same',
+                     kernel_regularizer=regularizers.l2(weight_decay)))
     model.add(Activation('elu'))
     model.add(BatchNormalization())
     model.add(MaxPooling2D(pool_size=(2, 2)))
@@ -218,14 +238,16 @@ def _create_random_cifar_model(input_shape):
 def _get_random_cifar_model(batch_size=32):
     opt_rms = keras.optimizers.rmsprop(lr=0.001, decay=1e-6)
     datagen = ImageDataGenerator(rotation_range=90,
-                                 width_shift_range=0.1, height_shift_range=0.1, horizontal_flip=True)
+                                 width_shift_range=0.1, height_shift_range=0.1,
+                                 horizontal_flip=True)
     x_train, y_train, x_val, y_val, x_test, y_test = _load_cifar_dataset()
     model = _create_random_cifar_model(input_shape=x_train.shape[1:])
     datagen.fit(x_train)
-    model.compile(loss='categorical_crossentropy', optimizer=opt_rms, metrics=['accuracy'])
+    model.compile(loss='categorical_crossentropy', optimizer=opt_rms,
+                  metrics=['accuracy'])
     model.fit_generator(datagen.flow(x_train, y_train, batch_size=batch_size),
-                        steps_per_epoch=x_train.shape[0] // batch_size, epochs=1,
-                        verbose=1, validation_data=(x_test, y_test))
+                        steps_per_epoch=x_train.shape[0] // batch_size,
+                        epochs=1, verbose=0, validation_data=(x_test, y_test))
 
     tb = datagen.flow(x_val, y_val, batch_size=batch_size)
     vb = datagen.flow(x_test, y_test, batch_size=batch_size)
@@ -247,7 +269,8 @@ def test_cifar10():
                      train_batches=tb3, val_batches=vb3)
 
     stack = StackEnsemble()
-    stack.model = RandomForestRegressor(verbose=1, n_estimators=1000, max_depth=7, n_jobs=4)
+    stack.model = RandomForestRegressor(verbose=0, n_estimators=3000,
+                                        max_depth=3, n_jobs=4)
     stack.add_member(member1)
     stack.add_member(member2)
     stack.add_member(member3)
